cleanup, more refactoring
