+++ api/app/Generators/DocGenerator.php
@@ -2,18 +2,53 @@
 
 namespace App\Generators;
 
+use Illuminate\Support\Facades\Log;
 use PhpOffice\PhpWord\Element;
 use PhpOffice\PhpWord\IOFactory;
 use PhpOffice\PhpWord\PhpWord;
 
-abstract class DocGenerator extends FileGenerator
+abstract class DocGenerator extends FileGenerator implements FileGeneratorInterface
 {
-    protected PhpWord $doc;
+    protected ?PhpWord $doc;
 
     protected array $strong;
 
-    public function __construct()
+    public function __construct(public string $fileName, protected ?string $dir)
     {
+        parent::__construct($fileName, $dir);
+    }
+
+    public function write()
+    {
+        if (! $this->doc) {
+            Log::error('Doc must be generated before writing.');
+
+            return;
+        }
+
+        try {
+
+            $path = $this->getPath($this->fileName, $this->dir);
+
+            $writer = IOFactory::createWriter($this->doc);
+
+            $writer->save($path);
+        } catch (\Exception $e) {
+            Log::error('Error saving doc: '.$this->fileName.' '.$e->getMessage());
+            throw $e;
+        }
+    }
+
+    protected function addLabelText(Element\Section $section, string $label, string $text)
+    {
+        $run = $section->addTextRun();
+        $run->addText($label.': ', $this->strong);
+        $run->addText($text);
+    }
+
+    protected function setup()
+    {
+
         $this->doc = new PhpWord();
 
         $this->doc->addTitleStyle(1, ['size' => 22, 'bold' => true], ['spaceBefore' => 240, 'spaceAfter' => 120]);
@@ -23,22 +58,5 @@ abstract class DocGenerator extends FileGenerator
         $this->doc->addTitleStyle(4, ['size' => 12, 'bold' => true], ['spaceBefore' => 240, 'spaceAfter' => 120]);
 
         $this->strong = ['bold' => true];
-    }
-
-    public function write(string $fileName, ?string $dir)
-    {
-
-        $path = $this->getPath($fileName, $dir);
-
-        $writer = IOFactory::createWriter($this->doc);
-
-        return $writer->save($path);
-    }
-
-    protected function addLabelText(Element\Section $section, string $label, string $text)
-    {
-        $run = $section->addTextRun();
-        $run->addText($label.': ', $this->strong);
-        $run->addText($text);
     }
 }
