+++ api/app/Generators/FileGeneratorInterface.php
@@ -7,4 +7,6 @@ interface FileGeneratorInterface
     public function generate(): self;
 
     public function write();
+
+    public function getFileName(): string;
 }
